backend/requirements.txt
skir-client >= 1.0.5
flask
flask-cors
